% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{restructure}
\alias{restructure}
\title{Restructure gcms postrun rawdata}
\usage{
restructure(filepath, delimiter = ",", write = FALSE)
}
\arguments{
\item{filepath}{the path to the raw data file}

\item{delimiter}{the delimiter of the raw data file to read in. Standard is set to comma.}

\item{write}{Whether the function should write the restructured datafile.
Value must be either TRUE or FALSE, false is standard.
If TRUE, the file is written into the path of the raw data file with
the extension "_tidied.csv".}
}
\value{
the restructured data frame.
}
\description{
Adds the name of sample as column next to the data columns
}
\details{
Function which gets the name of the sample from the row in the rawdata
and then adds it as the column "sample" as first column, allowing for
further data manipulation, e.g. pivoting, selection, filtering, etc
}
