% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vectorize_samplenames}
\alias{vectorize_samplenames}
\title{Title}
\usage{
vectorize_samplenames(filepath, allsamples_names, rownumber_tidydata)
}
\arguments{
\item{rownumber_tidydata}{}
}
\description{
Title
}
